+++ HISTORY.rst
@@ -1,6 +1,9 @@
 4.6.10 (XXXX-XX-XX)
 ===================
 - Fix base search class `_search` interface.
+- Add `SEARCH_QUERY_RESULTS_LIMIT_ERROR`. Control whether or not to raise
+  an error when the number of search results exceed the value of
+  `SEARCH_QUERY_RESULTS_LIMIT`. Default to `True` for backwards compatibility.
 
 4.6.9 (2025-08-31)
 ==================